@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -65,6 +65,8 @@ dependencies:
   sqflite_common_ffi: ^2.3.0+2
   pdf: ^3.10.4
   printing: ^5.11.0
+  cached_network_image: ^3.3.0
+  image_size_getter: ^2.1.2
 
 dependency_overrides:
   collection:
