@@ -1,21 +1,33 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:core';
 
 import 'package:dio/dio.dart';
+import 'package:moki_tutor/data/api/models/responses/subject_dto.dart';
 
 import 'package:moki_tutor/domain/interfaces/i_api.dart';
+
 import 'package:moki_tutor/domain/models/user.dart';
 import 'package:moki_tutor/domain/models/course.dart';
 
 import '../../domain/models/fragment.dart';
+import '../../domain/models/pdf_fragment.dart';
+import '../../domain/models/responses/paginated_sibjects.dart';
 import '../../domain/models/user_with_tokens.dart';
 import 'http_client/http_client.dart';
 import 'mapper/mapper.dart';
+import 'models/requests/fragment_request_data.dart';
 import 'models/requests/register_request.dart';
+import 'models/requests/request_add_pdf_subject.dart';
 import 'models/requests/request_add_subject.dart';
 import 'models/requests/request_confirmation.dart';
+import 'models/requests/request_delete_subject.dart';
+import 'models/requests/request_get_subject_fragments.dart';
+import 'models/requests/request_get_subjects.dart';
 import 'models/requests/request_login.dart';
-import 'models/responses/user_dto.dart';
+import 'models/requests/request_update_fragment.dart';
+import 'models/responses/pdf_fragment_dto.dart';
+
 import 'models/responses/user_with_tokens_dto.dart';
 
 typedef _Response = Response<Object?>?;
@@ -42,11 +54,15 @@ class Api implements IApi {
   }
 
   @override
-  Future<List<Course>?>? getPublishedCourses() async {}
+  Future<List<Course>?>? getPublishedCourses() async {
+    return null;
+  }
 
   @override
   Future<User?>? loginWithOtp(
-      {required String otp, required String email}) async {}
+      {required String otp, required String email}) async {
+    return null;
+  }
 
   @override
   Future<void> otpRequest(
@@ -78,7 +94,6 @@ class Api implements IApi {
         json.decode(response?.data as String) as Map<String, dynamic>));
   }
 
-  @override
   Future<void> addSubject(
       {required String pdfFile,
       required String title,
@@ -102,5 +117,66 @@ class Api implements IApi {
   @override
   Future<void> clearTokens() async {
     httpClient.clearTokens();
+  }
+
+  @override
+  Future<void> addPdfSubject(
+      {required String pdfFile,
+      required String title,
+      String? description,
+      required List<FragmentRequestData> fragments,
+      int? duration}) async {
+    await httpClient.request(RequestAddPdfSubject(
+        pdfFile: pdfFile,
+        title: title,
+        description: description,
+        fragments: fragments));
+  }
+
+  @override
+  Future<PaginatedSubjects> getSubjects(
+      {int limit = 50, int offset = 0}) async {
+    final result = await httpClient
+        .request(RequestGetSubjects(limit: limit, offset: offset));
+    print(result!.headers['X-Last-Row-Number']!.first);
+    print(result.headers['X-Total-Count']!.first);
+    return PaginatedSubjects(
+      offset: int.parse(result.headers['X-Last-Row-Number']!.first),
+      count: int.parse(result.headers['X-Total-Count']!.first),
+      subjects: (result.data as List).map(
+        (e) => mapper.mapSubject(SubjectDto.fromJson(e)),
+      ),
+    );
+  }
+
+  @override
+  Future<void> deleteSubject({required int id}) async {
+    await httpClient.request(RequestDeleteSubject(id: id));
+  }
+
+  @override
+  Future<List<PdfFragment>> getSubjectFragments({required int id}) async {
+    final result = await httpClient.request(RequestGetSubjectFragments(id: id));
+    return (jsonDecode(result!.data as String) as List)
+        .map((e) => mapper.mapPdfFragment(PdfFragmentDto.fromJson(e)))
+        .toList();
+  }
+
+  @override
+  Future<void> updateFragment({
+    required int id,
+    String? title,
+    String? description,
+    String? imagePath,
+    String? audioPath,
+    int? duration,
+  }) async {
+    await httpClient.request(RequestUpdateFragment(
+        id: id,
+        title: title,
+        description: description,
+        imagePath: imagePath,
+        audioPath: audioPath,
+        duration: duration));
   }
 }
