@@ -5,11 +5,11 @@ typedef DbData = Map<String, dynamic>;
 class LocalDbMapper {
   Subject mapSubject(DbData data) {
     return Subject(
-      id: data['id'],
-      title: data['title'],
-      description: data['description'],
-      date: DateTime.parse(data['created_at']),
-      pdfUrl: data['pdf_file'],
-    );
+        id: data['id'],
+        title: data['title'],
+        description: data['description'],
+        date: DateTime.parse(data['created_at']),
+        pdfUrl: data['pdf_file'],
+        countUncompleted: data['count_uncompleted']);
   }
 }
