@@ -1,6 +1,10 @@
+import 'dart:core';
 import 'dart:io';
 
+import '../../data/api/models/requests/fragment_request_data.dart';
 import '../models/course.dart';
+import '../models/pdf_fragment.dart';
+import '../models/responses/paginated_sibjects.dart';
 import '../models/user.dart';
 import '../models/fragment.dart';
 import '../models/user_with_tokens.dart';
@@ -33,11 +37,11 @@ abstract class IApi {
   Future<UserWithTokens> login(
       {required String email, required String password});
 
-  Future<void> addSubject(
+  Future<void> addPdfSubject(
       {required String pdfFile,
       required String title,
       String? description,
-      required DateTime date,
+      required List<FragmentRequestData> fragments,
       int? duration});
 
   // ---------------------------------------------------------------------------
@@ -45,4 +49,19 @@ abstract class IApi {
 
   // ---------------------------------------------------------------------------
   void setTokens(String access, String refresh);
+
+  Future<PaginatedSubjects> getSubjects({int limit = 50, int offset = 0});
+
+  Future<void> deleteSubject({required int id});
+
+  Future<List<PdfFragment>> getSubjectFragments({required int id});
+
+  Future<void> updateFragment({
+    required int id,
+    String? title,
+    String? description,
+    String? imagePath,
+    String? audioPath,
+    int? duration,
+  });
 }
