@@ -1,3 +1,4 @@
+import '../../data/api/models/requests/fragment_request_data.dart';
 import '../models/fragment.dart';
 import '../models/subject.dart';
 import '../models/subject_category.dart';
@@ -18,12 +19,14 @@ abstract class ISubjectsRepository {
     required String pdfFile,
     required String title,
     String? description,
-    required DateTime date,
+    required List<FragmentRequestData> fragments,
     int? duration,
   });
 
   Future<void> getSubjects();
   Future<int> updateSubject(Subject subject);
+
+  Future<void> deleteSubject(int id);
 
   void addChangeListener(Function() listener);
   void removeChangeListener(Function() listener);
