@@ -7,6 +7,6 @@ class PdfCreateSubjectEvent with _$PdfCreateSubjectEvent {
   const factory PdfCreateSubjectEvent.savePdfSubject(
       {required String title,
       required String pdfFile,
-      String? description,
-      required List<PdfFragment> pdfFragmentList}) = _EventSavePdfSubject;
+      required String description,
+      required List<PdfFragmentSample> pdfFragmentList}) = _EventSavePdfSubject;
 }
